Add github action to publish python packages
The plan moving forward for Elation is we will be using CodeArtifact as our
means to pip install packages. The existing devpi server will be too flaky for
us to use long term.

What this commit does is make a generic github action for us to be able to
publish said package to a CodeArtifact pypi repository. This will help us as we
start writing more re-usable code that can be shared across micro-services.

tickets:
https://elationhealth.atlassian.net/browse/RM-23184